The app, now, displays all continents and respective countries. Seems stable!

+++ app/src/main/java/app/covidstats/model/data/Continent.kt
@@ -12,7 +12,7 @@ data class Continent(
     val criticalPerOneMillion: Double,
     val deaths: Int,
     val deathsPerOneMillion: Double,
-    val population: Int,
+    val population: Long,
     val recovered: Int,
     val recoveredPerOneMillion: Double,
     val tests: Long,
