+++ app/src/main/java/app/covidstats/model/data/CountryInfo.kt
@@ -6,5 +6,5 @@ data class CountryInfo(
     val iso2: String,
     val iso3: String,
     val lat: Double,
-    val long: Int
+    val long: Double
 )
